feat: dateTime 함수 컴포넌트화
